fix: mark pre-query glob compile-DB match as provisional
Codex review, fresh evidence: _run_build_query resolves
sorted(sources.glob(cfg.compile_db)) AFTER the query has already run --
first lexically-sorted existing match wins. This module's own
dry-run preview of a glob-configured build.compile_db reported an
already-existing match as 'the resulting compile-DB path', with only
a vague 'the query may still recreate/refresh this file' caveat --
understating that the query can create a lexicographically-earlier
match (or remove the current one), making the real run resolve to a
genuinely DIFFERENT file than the one this preview reported.

Reworded to explicitly label the match as a pre-query snapshot
('resulting compile-DB path (provisional, pre-query snapshot): ...')
and spell out that the real run can select a different file.

New regression test:
test_glob_compile_db_pre_query_match_is_labeled_provisional (confirmed
to fail against the pre-fix code's unqualified wording).

Also splits tests/test_dry_run_contract.py's TestDumpDryRunBuildQueryTrust
class out into a new sibling file, tests/test_dry_run_build_query_contract.py
-- the original file crossed the AI-readiness 2000-line hard cap once this
fix's new test was added; this is purely a mechanical split (same tests,
same behavior), not a scope change.
